+++ ui_tests_playwright/automation/base/step/BaseDriverSteps.ts
@@ -6,8 +6,6 @@ import Buttons from '../../identifiers/Buttons';
 import {playwrightUtils} from '../../utils/PlaywrightUtils';
 import {urlsWithoutCookiesMessage} from '../../preconditionsData/UrlPreconditions';
 import UrlUtils from '../../utils/UrlUtils';
-import MainSiteButtons from '../../identifiers/mainSite/MainSiteButtons';
-import Calendly from '../../identifiers/mainSite/Calendly';
 
 class BaseDriverSteps {
 	public async createsNewBrowser(browserName: BrowsersEnum = BrowsersEnum.DEFAULT_BROWSER) {
@@ -207,42 +205,6 @@ class BaseDriverSteps {
 			document.documentElement.scrollTop = 0;
 			document.body.scrollTop = 0;
 		});
-	}
-
-	public async checkMemberCardCalendly(memberCard: Locator | undefined, memberData: {name: string; role: string}) {
-		if (!memberCard) {
-			throw new Error(`No card found for expert: ${memberData.name}`);
-		}
-		await expect(memberCard).toBeVisible();
-
-		const name = memberCard.getByTestId(Container.MemberName);
-		const role = memberCard.getByTestId(Container.MemberRole);
-		const consultButton = memberCard.getByTestId(MainSiteButtons.ScheduleAConsultationInCalendly);
-		const image = memberCard.locator('.member-foto');
-
-		await expect(name).toHaveText(memberData.name);
-		await expect(role).toHaveText(memberData.role);
-		await expect(consultButton).toBeVisible();
-		await expect(consultButton).toBeEnabled();
-		await this.checkImagesVisibility(image, 1);
-	}
-
-	public async findMatchingMemberCardByName(cardElements: Locator[], expertName: string) {
-		for (const card of cardElements) {
-			const cardName = await card.getByTestId(Container.MemberName).textContent();
-			if (cardName === expertName) return card;
-		}
-	}
-
-	public async checkAppropriateCalendlyModalOpensAndCloses(memberCard: Locator | undefined) {
-		if (!memberCard) {
-			throw new Error(`No card found`);
-		}
-		await memberCard.getByTestId(MainSiteButtons.ScheduleAConsultationInCalendly).click();
-		const calendlyFrame = driver.frameLocator(Calendly.frame);
-		await expect(calendlyFrame).toBeTruthy();
-		await driver.locator(Calendly.close).click();
-		await expect(driver.locator(Calendly.frame)).toBeHidden();
 	}
 }
